Move theme one level up.
