+++ src/App.js
@@ -4,7 +4,7 @@ import Button from '@material-ui/core/Button';
 import logo from './logo.svg';
 import './App.css';
 
-import theme from './components/theme';
+import theme from './theme';
 
 function App() {
   return (
