+++ src/Provider.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { ThemeProvider } from '@material-ui/styles';
-import theme from './components/theme'
+import theme from './theme'
 
 export default function Provider({ children }) {
   return <ThemeProvider theme={theme}>{children}</ThemeProvider>
